Convert the external API to use BufReader
Now it shouldn't matter whether the user wants to pass in 'raw' files or
read their data themself and then pass the data - both are supported by
BufReader, although it's less ergonomic than passing in byte data.

It might also be a decent idea to convert internal logic to read data in
a buffered manner as well, but that's for another day. At least the
API's roughly where it should be now.